requirements.txt
streamlit>=1.44.0
requests>=2.32.0
python-dotenv>=1.0.1
pydantic>=2.10.0
SQLAlchemy>=2.0.36
pypdf>=5.1.0
openai>=1.58.1
anthropic>=0.42.0
pandas>=2.2.3
